# Further optimize `IODeferred`