--- a/core/shared/src/main/scala/cats/effect/IODeferred.scala
+++ b/core/shared/src/main/scala/cats/effect/IODeferred.scala
@@ -19,14 +19,36 @@ package cats.effect
 import java.util.concurrent.atomic.{AtomicInteger, AtomicReference}
 
 private final class IODeferred[A] extends Deferred[IO, A] {
-  import IODeferred.Sentinel
 
-  private[this] val cell = new AtomicReference[AnyRef](Sentinel)
+  private[this] val initial: IO[A] = IO.asyncCheckAttempt { cb =>
+    IO {
+      val stack = callbacks.push(cb)
+      val handle = stack.currentHandle()
+
+      def clear(): Unit = {
+        stack.clearCurrent(handle)
+        val clearCount = clearCounter.incrementAndGet()
+        if ((clearCount & (clearCount - 1)) == 0) // power of 2
+          clearCounter.addAndGet(-callbacks.pack(clearCount))
+        ()
+      }
+
+      val back = cell.get()
+      if (back eq initial) {
+        Left(Some(IO(clear())))
+      } else {
+        clear()
+        Right(back.asInstanceOf[IO.Pure[A]].value)
+      }
+    }
+  }
+
+  private[this] val cell = new AtomicReference(initial)
   private[this] val callbacks = CallbackStack[Right[Nothing, A]](null)
   private[this] val clearCounter = new AtomicInteger
 
   def complete(a: A): IO[Boolean] = IO {
-    if (cell.compareAndSet(Sentinel, a.asInstanceOf[AnyRef])) {
+    if (cell.compareAndSet(initial, IO.pure(a))) {
       val _ = callbacks(Right(a), false)
       callbacks.clear() // avoid leaks
       true
@@ -35,44 +57,13 @@ private final class IODeferred[A] extends Deferred[IO, A] {
     }
   }
 
-  def get: IO[A] = IO defer {
-    val back = cell.get()
-
-    if (back eq Sentinel) IO.asyncCheckAttempt { cb =>
-      IO {
-        val stack = callbacks.push(cb)
-        val handle = stack.currentHandle()
-
-        def clear(): Unit = {
-          stack.clearCurrent(handle)
-          val clearCount = clearCounter.incrementAndGet()
-          if ((clearCount & (clearCount - 1)) == 0) // power of 2
-            clearCounter.addAndGet(-callbacks.pack(clearCount))
-          ()
-        }
-
-        val back = cell.get()
-        if (back eq Sentinel) {
-          Left(Some(IO(clear())))
-        } else {
-          clear()
-          Right(back.asInstanceOf[A])
-        }
-      }
-    }
-    else
-      IO.pure(back.asInstanceOf[A])
-  }
+  def get: IO[A] = cell.get()
 
   def tryGet: IO[Option[A]] = IO {
     val back = cell.get()
-    if (back eq Sentinel)
+    if (back eq initial)
       None
     else
-      Some(back.asInstanceOf[A])
+      Some(back.asInstanceOf[IO.Pure[A]].value)
   }
-}
-
-private object IODeferred {
-  private val Sentinel = new AnyRef
 }