--- a/core/shared/src/main/scala/cats/effect/IODeferred.scala
+++ b/core/shared/src/main/scala/cats/effect/IODeferred.scala
@@ -20,26 +20,36 @@ import java.util.concurrent.atomic.{AtomicInteger, AtomicReference}
 
 private final class IODeferred[A] extends Deferred[IO, A] {
 
-  private[this] val initial: IO[A] = IO.asyncCheckAttempt { cb =>
-    IO {
-      val stack = callbacks.push(cb)
-      val handle = stack.currentHandle()
+  private[this] val initial: IO[A] = {
+    val await = IO.asyncCheckAttempt[A] { cb =>
+      IO {
+        val stack = callbacks.push(cb)
+        val handle = stack.currentHandle()
 
-      def clear(): Unit = {
-        stack.clearCurrent(handle)
-        val clearCount = clearCounter.incrementAndGet()
-        if ((clearCount & (clearCount - 1)) == 0) // power of 2
-          clearCounter.addAndGet(-callbacks.pack(clearCount))
-        ()
+        def clear(): Unit = {
+          stack.clearCurrent(handle)
+          val clearCount = clearCounter.incrementAndGet()
+          if ((clearCount & (clearCount - 1)) == 0) // power of 2
+            clearCounter.addAndGet(-callbacks.pack(clearCount))
+          ()
+        }
+
+        val back = cell.get()
+        if (back eq initial) {
+          Left(Some(IO(clear())))
+        } else {
+          clear()
+          Right(back.asInstanceOf[IO.Pure[A]].value)
+        }
       }
+    }
 
+    IO.defer {
       val back = cell.get()
-      if (back eq initial) {
-        Left(Some(IO(clear())))
-      } else {
-        clear()
-        Right(back.asInstanceOf[IO.Pure[A]].value)
-      }
+      if (back eq initial)
+        await
+      else
+        back.asInstanceOf[IO.Pure[A]]
     }
   }