Just getting it into the record, unfortunately benchmarks showed that it made no difference at best, or slower at worst. The idea was to try to cache the `get: IO[A]` value in the atomic reference, instead of `A`, similar to what `IOFiber` does for `join`.

<details>

after
```
[info] Benchmark                   (count)   Mode  Cnt     Score   Error  Units
[info] DeferredBenchmark.cancel         10  thrpt   20    39.504 ± 1.192  ops/s
[info] DeferredBenchmark.cancel        100  thrpt   20     4.421 ± 0.127  ops/s
[info] DeferredBenchmark.cancel       1000  thrpt   20     0.401 ± 0.018  ops/s
[info] DeferredBenchmark.complete       10  thrpt   20    30.081 ± 3.468  ops/s
[info] DeferredBenchmark.complete      100  thrpt   20     7.963 ± 1.865  ops/s
[info] DeferredBenchmark.complete     1000  thrpt   20     1.374 ± 0.194  ops/s
[info] DeferredBenchmark.get            10  thrpt   20  1050.021 ± 9.871  ops/s
[info] DeferredBenchmark.get           100  thrpt   20   163.874 ± 1.474  ops/s
[info] DeferredBenchmark.get          1000  thrpt   20    16.155 ± 0.368  ops/s
```

before
```
[info] DeferredBenchmark.cancel         10  thrpt   20   42.230 ±  1.265  ops/s
[info] DeferredBenchmark.cancel        100  thrpt   20    4.343 ±  0.138  ops/s
[info] DeferredBenchmark.cancel       1000  thrpt   20    0.407 ±  0.016  ops/s
[info] DeferredBenchmark.complete       10  thrpt   20   31.948 ±  3.528  ops/s
[info] DeferredBenchmark.complete      100  thrpt   20    8.625 ±  1.620  ops/s
[info] DeferredBenchmark.complete     1000  thrpt   20    1.292 ±  0.267  ops/s
[info] DeferredBenchmark.get            10  thrpt   20  919.962 ± 90.952  ops/s
[info] DeferredBenchmark.get           100  thrpt   20  160.503 ±  1.369  ops/s
[info] DeferredBenchmark.get          1000  thrpt   20   20.408 ±  0.176  ops/s
```

</details>