@armanbilge 2023-04-17
Huh 🤔 `complete` got faster but `get` got _much_ slower and that seems bonkers. This PR replaces an `IO.defer(...)` a volatile read of an `IO.pure(...)`.

#### This PR
```
Benchmark                   (count)   Mode  Cnt    Score    Error  Units
DeferredBenchmark.cancel         10  thrpt   20   44.488 ±  2.784  ops/s
DeferredBenchmark.cancel        100  thrpt   20    5.303 ±  0.117  ops/s
DeferredBenchmark.cancel       1000  thrpt   20    0.449 ±  0.018  ops/s
DeferredBenchmark.complete       10  thrpt   20   32.755 ±  5.155  ops/s
DeferredBenchmark.complete      100  thrpt   20    8.012 ±  1.811  ops/s
DeferredBenchmark.complete     1000  thrpt   20    1.367 ±  0.044  ops/s
DeferredBenchmark.get            10  thrpt   20  702.181 ± 16.228  ops/s
DeferredBenchmark.get           100  thrpt   20   74.141 ±  3.685  ops/s
DeferredBenchmark.get          1000  thrpt   20    7.310 ±  0.139  ops/s
```

#### series/3.x
```
Benchmark                   (count)   Mode  Cnt     Score    Error  Units
DeferredBenchmark.cancel         10  thrpt   20    44.243 ±  1.105  ops/s
DeferredBenchmark.cancel        100  thrpt   20     4.598 ±  0.179  ops/s
DeferredBenchmark.cancel       1000  thrpt   20     0.433 ±  0.025  ops/s
DeferredBenchmark.complete       10  thrpt   20    22.119 ±  7.781  ops/s
DeferredBenchmark.complete      100  thrpt   20     7.368 ±  1.470  ops/s
DeferredBenchmark.complete     1000  thrpt   20     0.970 ±  0.284  ops/s
DeferredBenchmark.get            10  thrpt   20  1639.649 ± 60.727  ops/s
DeferredBenchmark.get           100  thrpt   20   258.978 ± 12.768  ops/s
DeferredBenchmark.get          1000  thrpt   20    28.177 ±  0.913  ops/s
```