@armanbilge 2023-04-17
There we go :)

```
Benchmark                    (count)   Mode  Cnt     Score     Error  Units
DeferredBenchmark.cancel          10  thrpt   20    47.890 ±   0.389  ops/s
DeferredBenchmark.cancel         100  thrpt   20     4.971 ±   0.110  ops/s
DeferredBenchmark.complete        10  thrpt   20    30.252 ±   5.486  ops/s
DeferredBenchmark.complete       100  thrpt   20     7.318 ±   2.162  ops/s
DeferredBenchmark.getAfter        10  thrpt   20  3532.888 ± 101.376  ops/s
DeferredBenchmark.getAfter       100  thrpt   20   985.239 ±  19.928  ops/s
DeferredBenchmark.getBefore       10  thrpt   20  1753.401 ±  17.655  ops/s
DeferredBenchmark.getBefore      100  thrpt   20   294.561 ±  13.756  ops/s
```