@armanbilge 2023-04-18
Good question, technically no, but there are some existing conflicts between the branches due to `asyncCheckAttempt` that we've hit before, and I didn't feel like flaring them up again /shrug. Also 3.5.x is around the corner ... right? 😛